refactor: create vote service
