+++ src/entities/Proposal/jobs.ts
@@ -1,11 +1,11 @@
 import JobContext from 'decentraland-gatsby/dist/entities/Job/context'
 import snakeCase from 'lodash/snakeCase'
 
-import { commentProposalUpdateInDiscourse } from '../../back/routes/proposal'
 import BidService from '../../services/BidService'
 import { BudgetService } from '../../services/BudgetService'
 import { DiscordService } from '../../services/DiscordService'
 import { ErrorService } from '../../services/ErrorService'
+import { ProposalService } from '../../services/ProposalService'
 import { ErrorCategory } from '../../utils/errorCategories'
 import { Budget } from '../Budget/types'
 
@@ -212,7 +212,7 @@ export async function finishProposal(context: JobContext) {
   const proposals = [...finishedProposals, ...acceptedProposals, ...rejectedProposals]
   context.log(`Updating ${proposals.length} proposals in discourse... \n\n`)
   for (const { id, title, winnerChoice, outcomeStatus } of proposals) {
-    commentProposalUpdateInDiscourse(id)
+    ProposalService.commentProposalUpdateInDiscourse(id)
     if (outcomeStatus) {
       DiscordService.finishProposal(
         id,
