+++ src/services/ProposalService.ts
@@ -1,17 +1,21 @@
+import logger from 'decentraland-gatsby/dist/entities/Development/logger'
 import RequestError from 'decentraland-gatsby/dist/entities/Route/error'
 import { v1 as uuid } from 'uuid'
 import isUUID from 'validator/lib/isUUID'
 
-import { DiscoursePost } from '../clients/Discourse'
+import { VoteService } from '../back/services/vote'
+import { Discourse, DiscourseComment, DiscoursePost } from '../clients/Discourse'
 import { SnapshotProposalContent } from '../clients/SnapshotGraphqlTypes'
 import CoauthorModel from '../entities/Coauthor/model'
 import isDAOCommittee from '../entities/Committee/isDAOCommittee'
 import ProposalModel from '../entities/Proposal/model'
 import * as templates from '../entities/Proposal/templates'
+import { getUpdateMessage } from '../entities/Proposal/templates/messages'
 import { ProposalAttributes, ProposalStatus, ProposalType } from '../entities/Proposal/types'
 import { isGrantProposalSubmitEnabled } from '../entities/Proposal/utils'
 import { SNAPSHOT_SPACE } from '../entities/Snapshot/constants'
 import VotesModel from '../entities/Votes/model'
+import { inBackground } from '../helpers'
 import { getProfile } from '../utils/Catalyst'
 import Time from '../utils/date/Time'
 import { getEnvironmentChainId } from '../utils/votes/utils'
@@ -211,5 +215,23 @@ export class ProposalService {
     }
 
     return ProposalModel.parse(proposal)
+  }
+
+  static commentProposalUpdateInDiscourse(id: string) {
+    inBackground(async () => {
+      const updatedProposal: ProposalAttributes | undefined = await ProposalModel.findOne<ProposalAttributes>({ id })
+      if (!updatedProposal) {
+        logger.error('Invalid proposal id for discourse update', { id: id })
+        return
+      }
+      const votes = await VoteService.getVotes(updatedProposal.id)
+      const updateMessage = getUpdateMessage(updatedProposal, votes)
+      const discourseComment: DiscourseComment = {
+        topic_id: updatedProposal.discourse_topic_id,
+        raw: updateMessage,
+        created_at: new Date().toJSON(),
+      }
+      await Discourse.get().commentOnPost(discourseComment)
+    })
   }
 }
